# Conflicts:
#   app_zoo/classic_control/cartpole/entry/cartpole_dqn_main.py
#   app_zoo/classic_control/cartpole/entry/cartpole_ppo_main.py
#   app_zoo/classic_control/pendulum/entry/pendulum_td3_main.py
#   app_zoo/multiagent_particle/entry/cooperative_navigation_qmix_main.py
#   nervex/config/config.py
#   nervex/entry/serial_entry.py
#   nervex/entry/serial_entry_irl.py
#   nervex/rl_utils/adder.py
#   nervex/utils/__init__.py
#   nervex/worker/collector/tests/speed_test/test_collector_profile.py

+++ app_zoo/classic_control/cartpole/envs/cartpole_env.py
@@ -44,7 +44,7 @@ class CartPoleEnv(BaseEnv):
             self._env.seed(self._seed)
         self._final_eval_reward = 0
         obs = self._env.reset()
-        obs = to_ndarray(obs)
+        obs = to_ndarray(obs).astype(np.float32)
         return obs
 
     def close(self) -> None:
@@ -65,7 +65,7 @@ class CartPoleEnv(BaseEnv):
         self._final_eval_reward += rew
         if done:
             info['final_eval_reward'] = self._final_eval_reward
-        obs = to_ndarray(obs)
+        obs = to_ndarray(obs).astype(np.float32)
         rew = to_ndarray([rew])  # wrapped to be transfered to a Tensor with shape (1,)
         return BaseEnvTimestep(obs, rew, done, info)
 
@@ -78,7 +78,7 @@ class CartPoleEnv(BaseEnv):
                 {
                     'min': [-4.8, float("-inf"), -0.42, float("-inf")],
                     'max': [4.8, float("inf"), 0.42, float("inf")],
-                    'dtype': float,
+                    'dtype': np.float32,
                 },
             ),
             # [min, max)
